fixing cat with CTRL-C

+++ execution/handle_signals.c
@@ -6,7 +6,7 @@
 /*   By: mmoumni <mmoumni@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/17 18:12:36 by mmoumni           #+#    #+#             */
-/*   Updated: 2022/07/20 23:01:45 by mmoumni          ###   ########.fr       */
+/*   Updated: 2022/07/21 09:22:21 by mmoumni          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,10 +16,8 @@
 
 void	sig_handler(int signal)
 {
-	if (signal == SIGINT && !g_minishell.herdoc
-		&& g_minishell.child)
-		;
-	else if (signal == SIGINT && g_minishell.herdoc)
+	if (signal == SIGINT && g_minishell.herdoc
+		&& !g_minishell.child)
 	{
 		kill(g_minishell.herdoc, SIGKILL);
 		g_minishell.herdoc = 0;
@@ -31,5 +29,11 @@ void	sig_handler(int signal)
 		rl_on_new_line();
 		rl_replace_line("", 0);
 		rl_redisplay();
+	}
+	else if (signal == SIGINT && !g_minishell.herdoc
+		&& g_minishell.child)
+	{
+		kill(g_minishell.child, SIGKILL);
+		g_minishell.child = 0;
 	}
 }
